+++ validation/kubernetes/eks-core/main.tf
@@ -33,17 +33,11 @@ module "eks_core" {
     role_arn = "foobar"
   }
   fluxcd_v2_config = {
-    type = "github"
     github = {
       org             = ""
       app_id          = 0
       installation_id = 0
       private_key     = ""
-    }
-    azure_devops = {
-      pat  = ""
-      org  = ""
-      proj = ""
     }
   }